fix(writer-guard): show access denied toast on SSR full-page-load
writerGuard previously called MessageService.add() directly, which has no
effect when the guard runs server-side under RenderMode.Server (all routes).
Copy-pasting a write route URL produced a silent redirect with no feedback.

Fix: encode the denial reason as _notice=<writeFeature> in the redirect URL
instead of calling MessageService directly. AppComponent subscribes to
NavigationEnd, detects _notice, shows the contextual "Access Denied" toast
via MessageService, then strips the param from the URL via
Location.replaceState (no second navigation cycle).

Works for both SPA navigation (NavigationEnd fires after guard redirect) and
full-page-load (Angular initial navigation fires NavigationEnd after hydration,
_notice is already in the server-delivered URL). ACCESS_DENIED_MESSAGES covers
all writerGuard writeFeature values: meetings, mailing-lists, votes, surveys,
committees, plus a generic fallback for unknown keys.
