+++ apps/lfx-one/src/app/app.component.ts
@@ -1,10 +1,13 @@
 // Copyright The Linux Foundation and each contributor to LFX.
 // SPDX-License-Identifier: MIT
 
+import { Location } from '@angular/common';
 import { Component, inject, makeStateKey, REQUEST_CONTEXT, TransferState } from '@angular/core';
-import { RouterOutlet } from '@angular/router';
+import { NavigationEnd, Router, RouterOutlet } from '@angular/router';
 import { AuthContext, User } from '@lfx-one/shared/interfaces';
+import { MessageService } from 'primeng/api';
 import { ToastModule } from 'primeng/toast';
+import { filter } from 'rxjs';
 
 import { getRuntimeConfig } from './shared/providers/runtime-config.provider';
 import { AccountContextService } from './shared/services/account-context.service';
@@ -14,6 +17,14 @@ import { IntercomService } from './shared/services/intercom.service';
 import { PlausibleService } from './shared/services/plausible.service';
 import { SegmentService } from './shared/services/segment.service';
 import { UserService } from './shared/services/user.service';
+
+const ACCESS_DENIED_MESSAGES: Record<string, string> = {
+  meetings: "You don't have permission to schedule meetings for this project.",
+  'mailing-lists': "You don't have permission to manage mailing lists for this project.",
+  votes: "You don't have permission to manage votes for this project.",
+  surveys: "You don't have permission to manage surveys for this project.",
+  committees: "You don't have permission to manage committees for this project.",
+};
 
 @Component({
   selector: 'lfx-root',
@@ -93,6 +104,8 @@ export class AppComponent {
       // Set DataDog RUM user context for session tracking
       this.dataDogRumService.setUser(this.auth.user);
     }
+
+    this.initAccessDeniedToast();
   }
 
   // Fails closed: missing JWT or App ID skips boot.
@@ -132,6 +145,34 @@ export class AppComponent {
       user_id: userId,
       name: user.name,
       email: user.email,
+    });
+  }
+
+  // Detects _notice query param placed by writerGuard on denial and shows the "Access
+  // Denied" toast. Using a URL param rather than calling MessageService directly in the
+  // guard is necessary because the guard runs server-side under RenderMode.Server —
+  // MessageService.add() on the server has no DOM to render into. The param survives the
+  // SSR redirect so the client always sees it on NavigationEnd regardless of how the user
+  // arrived (SPA click or copy-paste full-page-load).
+  private initAccessDeniedToast(): void {
+    const router = inject(Router);
+    const location = inject(Location);
+    const messageService = inject(MessageService);
+
+    router.events.pipe(filter((e) => e instanceof NavigationEnd)).subscribe(() => {
+      const parsed = router.parseUrl(router.url);
+      const notice = parsed.queryParams['_notice'];
+      if (!notice) return;
+
+      messageService.add({
+        severity: 'warn',
+        summary: 'Access Denied',
+        detail: ACCESS_DENIED_MESSAGES[notice] ?? "You don't have permission to perform this action for this project.",
+      });
+
+      // Remove _notice from the URL without triggering another navigation cycle
+      delete parsed.queryParams['_notice'];
+      location.replaceState(router.serializeUrl(parsed));
     });
   }
 }
