+++ apps/lfx-one/src/app/shared/guards/writer.guard.ts
@@ -1,23 +1,13 @@
 // Copyright The Linux Foundation and each contributor to LFX.
 // SPDX-License-Identifier: MIT
-
 import { inject } from '@angular/core';
 import { ActivatedRouteSnapshot, CanActivateFn, Router } from '@angular/router';
-import { MessageService } from 'primeng/api';
 import { catchError, map, of, switchMap } from 'rxjs';
 
 import { CommitteeService } from '../services/committee.service';
 import { PersonaService } from '../services/persona.service';
 import { ProjectContextService } from '../services/project-context.service';
 import { ProjectService } from '../services/project.service';
-
-const WRITE_FEATURE_MESSAGES: Record<string, string> = {
-  meetings: "You don't have permission to schedule meetings for this project.",
-  'mailing-lists': "You don't have permission to manage mailing lists for this project.",
-  votes: "You don't have permission to manage votes for this project.",
-  surveys: "You don't have permission to manage surveys for this project.",
-  committees: "You don't have permission to manage committees for this project.",
-};
 
 /**
  * Protects create/edit/admin routes that require project write permission.
@@ -38,17 +28,17 @@ const WRITE_FEATURE_MESSAGES: Record<string, string> = {
  * Redirects to the lens-appropriate overview on denial so the correct project context is
  * preserved and NavigationService.applyDefaultSelection does not override the selection.
  *
- * On denial — including when the project fetch fails with 403/404 (committee member with no
- * direct project-level OpenFGA relation) — shows a warning toast. Routes opt into a
- * contextual message by setting `data.writeFeature` (e.g. `'meetings'`, `'votes'`); falls
- * back to a generic message when absent.
+ * On denial, encodes `_notice=<writeFeature>` in the redirect URL instead of calling
+ * MessageService directly. AppComponent detects `_notice` on NavigationEnd, shows the
+ * contextual "Access Denied" toast, and strips the param via Location.replaceState. This
+ * two-step approach works for both SPA navigation and full-page-load (SSR) scenarios where
+ * MessageService.add() on the server has no client-side effect.
  */
 export const writerGuard: CanActivateFn = (route: ActivatedRouteSnapshot) => {
   const personaService = inject(PersonaService);
   const projectContextService = inject(ProjectContextService);
   const projectService = inject(ProjectService);
   const committeeService = inject(CommitteeService);
-  const messageService = inject(MessageService);
   const router = inject(Router);
 
   if (personaService.currentPersona() === 'executive-director') {
@@ -64,15 +54,10 @@ export const writerGuard: CanActivateFn = (route: ActivatedRouteSnapshot) => {
   if (!slug) {
     return router.parseUrl(overviewPath);
   }
-  const deniedUrl = router.createUrlTree([overviewPath], { queryParams: { project: slug } });
 
   const writeFeature: string | undefined = route.data?.['writeFeature'];
-  const deniedMessage = (writeFeature && WRITE_FEATURE_MESSAGES[writeFeature]) ?? "You don't have permission to perform this action for this project.";
-
-  const deny = () => {
-    messageService.add({ severity: 'warn', summary: 'Access Denied', detail: deniedMessage });
-    return deniedUrl;
-  };
+  const deniedUrl = router.createUrlTree([overviewPath], { queryParams: { project: slug, _notice: writeFeature ?? 'access' } });
+  const deny = () => deniedUrl;
 
   return projectService.getProject(slug, false, { meetingCoordinator: writeFeature === 'meetings' }).pipe(
     switchMap((project) => {
